Python-items/requirements.txt
anyascii==0.2.0
argcomplete==1.10.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
autocorrect==0.4.4
backcall==0.2.0
beautifulsoup4==4.8.0
bleach==3.3.0
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contractions==0.0.21
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
Flask==1.1.1
gensim==3.8.0
idna==2.8
IMAPClient==2.1.0
inflect==2.1.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.4.5
notebook==6.4.0
numpy==1.20.3
olefile==0.46
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
pdfminer.six==20181108
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pyahocorasick==1.4.2
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.4
Pygments==2.9.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
requests==2.22.0
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.7.1
six==1.12.0
sklearn==0.0
smart-open==3.0.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.6
SpeechRecognition==3.8.1
srsly==2.4.1
terminado==0.10.1
testpath==0.5.0
textract==1.6.3
textsearch==0.0.21
thinc==8.0.6
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typer==0.3.2
tzlocal==1.5.1
Unidecode==1.2.0
urllib3==1.25.11
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.4.3